+++ src/Service/FileDeleter/FileDeleter.php
@@ -2,6 +2,7 @@
 namespace Psys\OrderInvoiceBundle\Service\FileDeleter;
 
 use Doctrine\ORM\EntityManagerInterface;
+use Psys\OrderInvoiceBundle\Entity\InvoiceAdvance;
 use Psys\OrderInvoiceBundle\Entity\Order;
 use Psys\OrderInvoiceBundle\Model\Invoice\InvoiceType;
 use Symfony\Component\Filesystem\Filesystem;
@@ -27,6 +28,14 @@ class FileDeleter
     }
 
     /**
+     * Deletes the advance invoice file from disk and removes its reference from the database.
+     */
+    public function deleteAdvance(InvoiceAdvance $ent_InvoiceAdvance, ?string $nameFileSystem = null): void
+    {
+        $this->delete($ent_InvoiceAdvance->getInvoice()->getOrder(), InvoiceType::ADVANCE, $nameFileSystem, $ent_InvoiceAdvance);
+    }
+
+    /**
      * Deletes the final invoice file from disk and removes its reference from the database.
      */
     public function deleteFinal(Order $ent_Order, ?string $nameFileSystem = null): void
@@ -35,9 +44,17 @@ class FileDeleter
     }
 
     /**
+     * Deletes the regular invoice file from disk and removes its reference from the database.
+     */
+    public function deleteRegular(Order $ent_Order, ?string $nameFileSystem = null): void
+    {
+        $this->delete($ent_Order, InvoiceType::REGULAR, $nameFileSystem);
+    }
+
+    /**
      * Deletes the file from disk and removes its reference from the database.
      */
-    private function delete(Order $ent_Order, InvoiceType $invoiceType, ?string $nameFileSystem = null): void
+    private function delete(Order $ent_Order, InvoiceType $invoiceType, ?string $nameFileSystem = null, ?InvoiceAdvance $ent_InvoiceAdvance = null): void
     {        
         if ($invoiceType === InvoiceType::PROFORMA)
         {
@@ -45,11 +62,22 @@ class FileDeleter
             $ent_InvoiceProforma = $ent_Order->getInvoice()->getInvoiceProforma();
             $ent_File = $ent_InvoiceProforma->getFile();
         }
+        else if ($invoiceType === InvoiceType::ADVANCE)
+        {
+            $storagePath = $this->storagePath['advance'];
+            $ent_File = $ent_InvoiceAdvance->getFile();
+        }
         else if ($invoiceType === InvoiceType::FINAL)
         {
             $storagePath = $this->storagePath['final'];
             $ent_InvoiceFinal = $ent_Order->getInvoice()->getInvoiceFinal();
             $ent_File = $ent_InvoiceFinal->getFile();
+        }
+        else if ($invoiceType === InvoiceType::REGULAR)
+        {
+            $storagePath = $this->storagePath['regular'];
+            $ent_InvoiceRegular = $ent_Order->getInvoice()->getInvoiceRegular();
+            $ent_File = $ent_InvoiceRegular->getFile();
         }
 
         // No file to delete
@@ -66,10 +94,20 @@ class FileDeleter
             $ent_InvoiceProforma->setFile(null);
             $this->em->persist($ent_InvoiceProforma);
         }
+        else if ($invoiceType === InvoiceType::ADVANCE)
+        {
+            $ent_InvoiceAdvance->setFile(null);
+            $this->em->persist($ent_InvoiceAdvance);
+        }
         else if ($invoiceType === InvoiceType::FINAL)
         {
             $ent_InvoiceFinal->setFile(null);
             $this->em->persist($ent_InvoiceFinal);
+        }
+        else if ($invoiceType === InvoiceType::REGULAR)
+        {
+            $ent_InvoiceRegular->setFile(null);
+            $this->em->persist($ent_InvoiceRegular);
         }
         
         $this->em->remove($ent_File);